+++ labs/tasks/__init__.py
@@ -1,0 +1,32 @@
+from labs.tasks.llm import (
+    vectorize_repo_to_database_task,
+    find_similar_embeddings_task,
+    prepare_prompt_and_context_task,
+    get_llm_response_task,
+)
+from labs.tasks.repo import (
+    get_issue_task,
+    create_branch_task,
+    clone_repo_task,
+    apply_code_changes_task,
+    commit_changes_task,
+    create_pull_request_task,
+)
+from labs.tasks.run import init_task, run_on_repo_task, run_on_local_repo_task
+
+
+__all__ = [
+    "vectorize_repo_to_database_task",
+    "find_similar_embeddings_task",
+    "prepare_prompt_and_context_task",
+    "get_llm_response_task",
+    "get_issue_task",
+    "create_branch_task",
+    "clone_repo_task",
+    "apply_code_changes_task",
+    "commit_changes_task",
+    "create_pull_request_task",
+    "init_task",
+    "run_on_repo_task",
+    "run_on_local_repo_task",
+]
